[Issue 4] Went through to figure out how to register a script to the head of the page right above the title.
I will be using this to handle the autocompleter there so I can have
finer control over the javascript. I should be able to handle it in
external javascript files with the registerScriptFile but for now I am
using registerScript which takes a string that will be inserted into the
head.